refactor: deepen module architecture across 7 packages
Seven targeted refactors to turn shallow modules into deep ones:

1. Extract shared morphology helpers (applyCase, spanishPluralize)
   to providers/src/morphology.ts — deduplicated from 4 files
2. Remove hidden dependency: @dialectos/locale-utils no longer
   re-exports @dialectos/security symbols
3. Add compile-time completeness check for dialect metadata
   (all 25 dialects validated)
4. Remove 3 pass-through factory wrappers — CLI and MCP now
   import directly from @dialectos/providers
5. Extract inline data to JSON files: garbage patterns, false
   friends, accent fixes, word lists, gender overrides
6. Split llm.ts (955 lines) into 3 focused modules: llm.ts
   (class + wire protocol), llm-prompts.ts (prompt construction),
   llm-response.ts (response extraction + sanitization)
7. Reorganize @dialectos/types index with clear separation
   between pure types and runtime data modules

All 566 tests pass, 0 failures.
