+++ packages/mcp/src/__tests__/docs.test.ts
@@ -21,7 +21,7 @@ vi.mock("node:fs", async () => {
   return {
     ...actual,
     readFileSync: vi.fn((path: string | URL, ...args: any[]) => {
-      if (typeof path === "string" && (path.includes("dialectal-dictionary.json") || path.includes("verb-conjugations.json"))) {
+      if (typeof path === "string" && path.endsWith(".json")) {
         return actual.readFileSync(path, ...args);
       }
       return "# Hello World\n\nThis is a test paragraph.\n\n```javascript\nconsole.log('code');\n```";
@@ -75,12 +75,14 @@ vi.mock("@dialectos/providers", () => ({
       get: vi.fn(),
       getAuto: vi.fn(),
       register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
     };
   }),
   createProviderRegistry: vi.fn().mockReturnValue({
     get: vi.fn(),
     getAuto: vi.fn(),
     register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
   }),
   getDefaultProviderRegistry: vi.fn(),
   DeepLProvider: vi.fn(),
@@ -120,6 +122,7 @@ describe("MCP Docs Tools", () => {
       get: vi.fn(),
       getAuto: vi.fn(),
       register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
     } as unknown as ProviderRegistry;
 
     // Mock parseMarkdown
