+++ packages/mcp/src/__tests__/i18n.test.ts
@@ -22,7 +22,7 @@ vi.mock("node:fs", async () => {
   return {
     ...actual,
     readFileSync: vi.fn((path: string | URL, ...args: any[]) => {
-      if (typeof path === "string" && (path.includes("dialectal-dictionary.json") || path.includes("verb-conjugations.json"))) {
+      if (typeof path === "string" && path.endsWith(".json")) {
         return actual.readFileSync(path, ...args);
       }
       return "# Hello World";
@@ -90,12 +90,14 @@ vi.mock("@dialectos/providers", () => ({
       get: vi.fn(),
       getAuto: vi.fn(),
       register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
     };
   }),
   createProviderRegistry: vi.fn().mockReturnValue({
     get: vi.fn(),
     getAuto: vi.fn(),
     register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
   }),
   getDefaultProviderRegistry: vi.fn(),
   DeepLProvider: vi.fn(),
@@ -137,6 +139,7 @@ describe("MCP i18n Tools", () => {
       get: vi.fn(),
       getAuto: vi.fn(),
       register: vi.fn(),
+      prepareRequest: vi.fn().mockReturnValue({ sourceLang: "en", targetLang: "es-ES", options: {}, warnings: [] }),
     } as unknown as ProviderRegistry;
 
     // Mock provider translate
