+++ packages/providers/src/data/garbage-patterns.json
@@ -1,0 +1,34 @@
+[
+  { "pattern": "```" },
+  { "pattern": "^\\s*(translation|traducci[oó]n)\\s*:", "flags": "i" },
+  { "pattern": "\\bhere is (a |the )?translat", "flags": "i" },
+  { "pattern": "\\bhere('s| is) (a |the )?(translated|spanish)\\b", "flags": "i" },
+  { "pattern": "\\bbelow is (a |the )?translat", "flags": "i" },
+  { "pattern": "\\baqu[ií] (est[aá]|tienes) (la )?traducci[oó]n\\b", "flags": "i" },
+  { "pattern": "\\bdialect quality contract\\b", "flags": "i" },
+  { "pattern": "\\blexical ambiguity constraints\\b", "flags": "i" },
+  { "pattern": "\\bforbidden output\\b", "flags": "i" },
+  { "pattern": "\\btaboo policy\\b", "flags": "i" },
+  { "pattern": "\\bdo not translate literally\\b", "flags": "i" },
+  { "pattern": "\\bsure,? i can help", "flags": "i" },
+  { "pattern": "\\bokay,? i understand", "flags": "i" },
+  { "pattern": "\\blet'?s begin", "flags": "i" },
+  { "pattern": "\\bof (the |your )?(provided |given |original )?text\\b", "flags": "i" },
+  { "pattern": "^\\s*<<<\\s*$", "flags": "m" },
+  { "pattern": "^\\s*elote\\s*$", "flags": "i" },
+  { "pattern": "^\\s*mazorca\\s*$", "flags": "i" },
+  { "pattern": "^\\s*es-[a-z]{2}\\s*$", "flags": "i" },
+  { "pattern": "\\/no_think" },
+  { "pattern": "^\\s*voseo\\s*:", "flags": "i" },
+  { "pattern": "^\\s*la respuesta es\\s*:", "flags": "i" },
+  { "pattern": "\\bnice\\s+[a-z]+\\b", "flags": "i" },
+  { "pattern": "\\bgood\\s+[a-z]+\\b", "flags": "i" },
+  { "pattern": "\\bpuede\\s+decirlo\\s+en\\s+español\\b", "flags": "i" },
+  { "pattern": "\\bpuedes\\s+decirlo\\s+en\\s+español\\b", "flags": "i" },
+  { "pattern": "\\bdilo\\s+en\\s+español\\b", "flags": "i" },
+  { "pattern": "\\btraduce\\s+(esto|lo siguiente)\\b", "flags": "i" },
+  { "pattern": "^\\s*¡?Bienvenido!?\\s*$", "flags": "i" },
+  { "pattern": "\\bEntiendo\\.?\\s*Estoy listo\\b", "flags": "i" },
+  { "pattern": "\\bfrutill[ao]\\b.*\\bfrutill[ao]\\b", "flags": "i" },
+  { "pattern": "\\bcampero\\b", "flags": "i" }
+]
