+++ packages/providers/src/data/spanish-word-lists.json
@@ -1,0 +1,25 @@
+{
+  "commonNonNouns": [
+    "que", "de", "en", "y", "a", "es", "se", "no", "por", "con",
+    "para", "como", "pero", "su", "mas", "si", "yo", "el", "lo",
+    "le", "me", "te", "nos", "les", "ya", "esta", "ese", "aquel",
+    "todo", "cada", "otro", "mismo", "mucho", "poco", "tanto",
+    "donde", "cuando", "aqui", "alli", "hoy", "ayer", "manana",
+    "siempre", "nunca", "tambien", "muy", "bien", "mal", "puede",
+    "tiene", "hace", "sido", "visto", "dado", "hecho", "dicho",
+    "puesto", "primero", "segundo", "tercero", "ultimo", "nuevo",
+    "viejo", "grande", "pequeno", "bueno", "malo", "mejor", "peor",
+    "principal", "general", "total", "social", "debe", "quiere",
+    "sabe", "viene", "creo", "parece", "dice", "haz", "pon",
+    "anda", "ando", "iendo", "ado", "ido"
+  ],
+  "invariantNouns": [
+    "crisis", "lunes", "martes", "miercoles", "jueves", "viernes",
+    "paraguas", "tijeras", "gafas", "pijama", "especies", "series",
+    "analysis", "sis", "bus"
+  ],
+  "invariantAdjectives": [
+    "verde", "joven", "grande", "fuerte", "dulce", "salada",
+    "cortes", "modales", "simple", "inteligible"
+  ]
+}
