+++ packages/providers/src/quality-gates.ts
@@ -9,6 +9,7 @@
 
 import type { SpanishDialect } from "@dialectos/types";
 import { validateDialectCompliance } from "@dialectos/types";
+import { GARBAGE_PATTERNS, COMMON_ENGLISH_WORDS } from "./data/garbage-patterns.js";
 
 export type ModelTier = "tiny" | "small" | "medium" | "large";
 
@@ -99,43 +100,6 @@ export function dialectComplianceCheck(context: QualityGateContext): QualityGate
 // Garbage pattern detection (moved from LLMProvider)
 // ============================================================================
 
-const GARBAGE_PATTERNS = [
-  /```/,
-  /^\s*(translation|traducci[oó]n)\s*:/i,
-  /\bhere is (a |the )?translat/i,
-  /\bhere('s| is) (a |the )?(translated|spanish)\b/i,
-  /\bbelow is (a |the )?translat/i,
-  /\baqu[ií] (est[aá]|tienes) (la )?traducci[oó]n\b/i,
-  /\bdialect quality contract\b/i,
-  /\blexical ambiguity constraints\b/i,
-  /\bforbidden output\b/i,
-  /\btaboo policy\b/i,
-  /\bdo not translate literally\b/i,
-  /\bsure,? i can help/i,
-  /\bokay,? i understand/i,
-  /\blet'?s begin/i,
-  /\bof (the |your )?(provided |given |original )?text\b/i,
-  /^\s*<<<\s*$/m,
-  /^\s*elote\s*$/i,
-  /^\s*mazorca\s*$/i,
-  /^\s*es-[a-z]{2}\s*$/i,
-  /\/no_think/,
-  /^\s*voseo\s*:/i,
-  /^\s*la respuesta es\s*:/i,
-  /\bnice\s+[a-z]+\b/i,
-  /\bgood\s+[a-z]+\b/i,
-  /\bpuede\s+decirlo\s+en\s+español\b/i,
-  /\bpuedes\s+decirlo\s+en\s+español\b/i,
-  /\bdilo\s+en\s+español\b/i,
-  /\btraduce\s+(esto|lo siguiente)\b/i,
-  /^\s*¡?Bienvenido!?\s*$/i,
-  /\bEntiendo\.?\s*Estoy listo\b/i,
-  /\bfrutill[ao]\b.*\bfrutill[ao]\b/i,
-  /\bcampero\b/i,
-];
-
-const COMMON_ENGLISH_WORDS =
-  /\b(the|is|are|was|were|have|has|had|do|does|did|will|would|could|should|may|might|can|this|that|these|those|with|from|into|through|during|before|after|above|below|between|under|again|further|then|once|here|there|when|where|why|how|all|any|both|each|few|more|most|other|some|such|no|nor|not|only|own|same|so|than|too|very|just|now|also|back|down|off|over|out|up|about|because|but|if|or|since|though|while|although|unless|until|whether|either|neither|both|and|yet|still|however|therefore|moreover|furthermore|nevertheless|otherwise|meanwhile|instead|besides|actually|probably|certainly|definitely|absolutely|completely|totally|exactly|precisely|specifically|particularly|especially|generally|usually|normally|typically|frequently|often|sometimes|occasionally|rarely|seldom|never|always|constantly|continuously|repeatedly|regularly|daily|weekly|monthly|yearly|early|late|soon|recently|already|yet|still|before|after|later|earlier|formerly|previously|currently|presently|immediately|instantly|directly|straight|slowly|quickly|rapidly|suddenly|gradually|eventually|finally|initially|originally|primarily|mainly|mostly|largely|partly|slightly|somewhat|fairly|pretty|rather|quite|very|extremely|incredibly|unbelievably|amazingly|surprisingly|remarkably|notably|significantly|substantially|considerably|greatly|deeply|strongly|weakly|hardly|barely|scarcely|nearly|almost|practically|virtually|essentially|basically|fundamentally|ultimately|absolutely|relatively|comparatively|exceptionally|extraordinarily|tremendously|enormously|hugely|vastly|widely|narrowly|closely|loosely|tightly|firmly|softly|gently|roughly|smoothly|easily|difficultly|simply|complexly|plainly|clearly|obviously|evidently|apparently|seemingly|presumably|supposedly|allegedly|reportedly|supposedly|theoretically|hypothetically|potentially|possibly|perhaps|maybe|likely|probably|presumably|undoubtedly|unquestionably|indisputably|incontrovertibly|indefinitely|permanently|temporarily|briefly|shortly)\b/gi;
 
 /**
  * Garbage pattern check: detect empty output, unchanged source, garbage patterns,
