+++ packages/types/src/gender-overrides.json
@@ -1,0 +1,168 @@
+{
+  "computadora": "f",
+  "ordenador": "m",
+  "computador": "m",
+  "celular": "m",
+  "movil": "m",
+  "telefono": "m",
+  "aplicacion": "f",
+  "pagina": "f",
+  "archivo": "m",
+  "carpeta": "f",
+  "ventana": "f",
+  "pantalla": "f",
+  "teclado": "m",
+  "raton": "m",
+  "impresora": "f",
+  "conexion": "f",
+  "contrasena": "f",
+  "cuenta": "f",
+  "servidor": "m",
+  "base de datos": "f",
+  "codigo": "m",
+  "enlace": "m",
+  "correo": "m",
+  "mensaje": "m",
+  "notificacion": "f",
+  "configuracion": "f",
+  "ajuste": "m",
+  "carro": "m",
+  "coche": "m",
+  "auto": "m",
+  "vehiculo": "m",
+  "autobus": "m",
+  "colectivo": "m",
+  "micro": "m",
+  "camion": "m",
+  "bicicleta": "f",
+  "motocicleta": "f",
+  "avion": "m",
+  "tren": "m",
+  "barco": "m",
+  "metro": "m",
+  "estacion": "f",
+  "parada": "f",
+  "boleto": "m",
+  "billete": "m",
+  "pasaje": "m",
+  "viaje": "m",
+  "ruta": "f",
+  "calle": "f",
+  "avenida": "f",
+  "carretera": "f",
+  "autopista": "f",
+  "zumo": "m",
+  "jugo": "m",
+  "agua": "f",
+  "cafe": "m",
+  "cerveza": "f",
+  "vino": "m",
+  "refresco": "m",
+  "bebida": "f",
+  "tarta": "f",
+  "pastel": "m",
+  "torta": "f",
+  "bizcocho": "m",
+  "fruta": "f",
+  "carne": "f",
+  "pescado": "m",
+  "pollo": "m",
+  "ensalada": "f",
+  "sopa": "f",
+  "pan": "m",
+  "arroz": "m",
+  "huevo": "m",
+  "leche": "f",
+  "manteca": "f",
+  "mantequilla": "f",
+  "verdura": "f",
+  "legumbre": "f",
+  "postre": "m",
+  "cabeza": "f",
+  "brazo": "m",
+  "pierna": "f",
+  "mano": "f",
+  "pie": "m",
+  "ojo": "m",
+  "boca": "f",
+  "nariz": "f",
+  "oreja": "f",
+  "cuello": "m",
+  "espalda": "f",
+  "pecho": "m",
+  "corazon": "m",
+  "estomago": "m",
+  "hueso": "m",
+  "sangre": "f",
+  "pelle": "f",
+  "cara": "f",
+  "diente": "m",
+  "lengua": "f",
+  "hombre": "m",
+  "mujer": "f",
+  "nino": "m",
+  "nina": "f",
+  "bebe": "m",
+  "madre": "f",
+  "padre": "m",
+  "hijo": "m",
+  "hija": "f",
+  "hermano": "m",
+  "hermana": "f",
+  "abuelo": "m",
+  "abuela": "f",
+  "tio": "m",
+  "tia": "f",
+  "primo": "m",
+  "prima": "f",
+  "esposo": "m",
+  "esposa": "f",
+  "amigo": "m",
+  "amiga": "f",
+  "persona": "f",
+  "gente": "f",
+  "arbol": "m",
+  "flor": "f",
+  "rio": "m",
+  "mar": "m",
+  "montana": "f",
+  "selva": "f",
+  "bosque": "m",
+  "playa": "f",
+  "sol": "m",
+  "luna": "f",
+  "estrella": "f",
+  "cielo": "m",
+  "tierra": "f",
+  "piedra": "f",
+  "hoja": "f",
+  "planta": "f",
+  "camisa": "f",
+  "pantalon": "m",
+  "vestido": "m",
+  "falda": "f",
+  "zapato": "m",
+  "sombrero": "m",
+  "chaqueta": "f",
+  "abrigo": "m",
+  "bufanda": "f",
+  "guante": "m",
+  "calcetin": "m",
+  "corbata": "f",
+  "casa": "f",
+  "puerta": "f",
+  "mesa": "f",
+  "silla": "f",
+  "cama": "f",
+  "sofa": "m",
+  "cocina": "f",
+  "bano": "m",
+  "habitacion": "f",
+  "cuarto": "m",
+  "escalera": "f",
+  "techo": "m",
+  "pared": "f",
+  "ventana": "f",
+  "llave": "f",
+  "espejo": "m"
+}
